Rent_a_friend/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
python-jose
